backend/pyproject.toml
[project]
name = "colreg-assistant"
version = "0.1.0"
description = "COLREG Assistant - Maritime navigation chatbot with RAG"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.2.0",
    "litellm>=1.55.0",
    "langgraph>=0.2.59",
    "langchain-core>=0.3.29",
    "langchain-text-splitters>=0.3.4",
    "supabase>=2.12.0",
    "vecs>=0.4.3",
    "google-genai>=0.2.2",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.3",
    "pymupdf>=1.25.0",
]

# Vercel deployment: points to FastAPI app
[project.scripts]
app = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
